0.8.1b Release
Changes:

Added non-blind injection support. This means that brute force isn't required if the hole isn't blind, and speeds the exploit up a lot.

Before determining "UNION SELECT" exploits aren't possible, sqlier now attempts to bypass some common filtering mechanisms before giving up.

Added support for subquery based exploits. This allows sqlier to fall back on a feature in MySQL 4+ that is harder to filter for the brute forcing.

Altered detection engine a lot, fixing a bunch of injection issues.

Changed the format of the saved exploits, putting the exploit URL in with the encoded part so that one string can represent the whole exploit.

wget options in the saved exploits are now encoded, fixing a bug where the wget option has a colon in it.

Fixed a bug on how the strings were encoded, creating spaces in the string which were problematic.

Various other bug fixes and engine tweaks.